+++ src/harlequin/options.py
@@ -1,45 +1,132 @@
 from __future__ import annotations
 
+import re
 from abc import ABC, abstractmethod
 from pathlib import Path
-from typing import Any, Callable
+from typing import Any, Callable, Generator, Sequence
 
 import click
+from textual.validation import ValidationResult, Validator
+from textual.widget import Widget
+
+from harlequin.copy_widgets import (
+    Input,
+    NoFocusLabel,
+    PathInput,
+    Select,
+    Switch,
+)
+
+
+class _CustomValidator(Validator):
+    def __init__(
+        self,
+        validator: Callable[[str], tuple[bool, str | None]] | None = None,
+        failure_description: str | None = None,
+    ) -> None:
+        super().__init__(failure_description)
+        self.validator = validator or (lambda _: (True, ""))
+
+    def validate(self, value: str) -> ValidationResult:
+        try:
+            is_valid, message = self.validator(value)
+        except Exception as e:
+            return self.failure(str(e))
+
+        if is_valid:
+            return self.success()
+        else:
+            return self.failure(message or "Validation failed.")
 
 
 class AbstractOption(ABC):
+    """
+    The ABC for Harlequin options that are used as both command-line options and
+    GUI options. Options have names and descriptions, and may have user-faacing
+    labels and aliased, short declarations (for CLI options).
+
+    Subclasses define options for specific data types, like text or boolean options.
+    """
+
     def __init__(
         self,
         name: str,
         description: str,
         *args: Any,
         label: str | None = None,
-        short_decls: list[str] | None = None,
+        short_decls: Sequence[str] | None = None,
         **kwargs: Any,
     ) -> None:
-        if short_decls is None:
-            short_decls = []
-        self.name = name
+        """
+        Args:
+            name (str): A unique name for this option. Must be a valid
+                HTML/CSS id and a valid CLI option name (without the `--` prefix).
+                e.g., "port", "header"
+            description (str): Help text for this option.
+            label (str | None): For GUI options, a human-friendly label for this option.
+            short_decls (Sequence[str] | None): For CLI options, a list of short aliases
+                (including the `-` prefix) for this option (e.g., ["-p"]).
+        """
+        # names should be valid html/css ids
+        if re.match(r"[A-Za-z](\w|-)*", name):
+            self.name = name
+        else:
+            raise ValueError(
+                "An Option's name attribute must match "
+                """r'[A-Za-z](\w|-)*' """
+                "so it is a valid CLI flag and HTML/CSS id."
+            )
         self.description = description
-        self.label = label
-        self.short_decls = short_decls
+        self.label = label or name.replace("_", " ").replace("_", " ").capitalize()
+        short_decls = short_decls or []
+        self.short_decls = [
+            decl if decl.startswith("-") else f"-{decl}" for decl in short_decls
+        ]
 
     @abstractmethod
     def to_click(self) -> Callable[[click.Command], click.Command]:
         pass
 
+    @abstractmethod
+    def to_widgets(self) -> Generator[Widget, None, None]:
+        pass
+
 
 class TextOption(AbstractOption):
+    """
+    An option for free text input, including optional validation.
+    """
+
     def __init__(
         self,
         name: str,
         description: str,
         label: str | None = None,
         short_decls: list[str] | None = None,
-        validator: Callable[[str], str] | None = None,
+        default: str | None = None,
+        placeholder: str | None = None,
+        validator: Callable[[str], tuple[bool, str | None]] | None = None,
     ) -> None:
+        """
+        Args:
+            name (str): A unique name for this option. Must be a valid
+                HTML/CSS id and a valid CLI option name (without the `--` prefix).
+                e.g., "port", "header"
+            description (str): Help text for this option.
+            label (str | None): For GUI options, a human-friendly label for this option.
+            short_decls (Sequence[str] | None): For CLI options, a list of short aliases
+                (including the `-` prefix) for this option (e.g., ["-p"]).
+            default (str | None): The default value for this option.
+            placeholder (str | None): For GUI options, placeholder text for this option.
+            validator (Callable[[str], tuple[bool, str | None]] | None): A callable that
+                receives the raw input as a string returns a tuple. The first item of
+                the tuple is either True for valid input or False for invalid input.
+                The second item is a message shown to the user if the validation fails.
+        """
         super().__init__(name, description, label=label, short_decls=short_decls)
         self.validator = validator
+        self.default = default
+        self.placeholder = placeholder
 
     def to_click(self) -> Callable[[click.Command], click.Command]:
         def click_callback(
@@ -47,9 +134,11 @@ class TextOption(AbstractOption):
         ) -> str:
             if self.validator is not None:
                 try:
-                    value = self.validator(value)
+                    is_valid, message = self.validator(value)
                 except Exception as e:
                     raise click.BadParameter(str(e)) from e
+                if not is_valid:
+                    raise click.BadParameter(message or "Validation failed.")
             return value
 
         return click.option(
@@ -57,6 +146,15 @@ class TextOption(AbstractOption):
             *self.short_decls,
             help=self.description,
             callback=click_callback,
+        )
+
+    def to_widgets(self) -> Generator[Widget, None, None]:
+        yield NoFocusLabel(f"{self.label}:", classes="input_label")
+        yield Input(
+            value=self.default or "",
+            placeholder=self.placeholder or "",
+            id=self.name,
+            validators=[_CustomValidator(self.validator)],
         )
 
 
@@ -68,6 +166,16 @@ class ListOption(AbstractOption):
         label: str | None = None,
         short_decls: list[str] | None = None,
     ) -> None:
+        """
+        Args:
+            name (str): A unique name for this option. Must be a valid
+                HTML/CSS id and a valid CLI option name (without the `--` prefix).
+                e.g., "port", "header"
+            description (str): Help text for this option.
+            label (str | None): For GUI options, a human-friendly label for this option.
+            short_decls (Sequence[str] | None): For CLI options, a list of short aliases
+                (including the `-` prefix) for this option (e.g., ["-p"]).
+        """
         super().__init__(name, description, label=label, short_decls=short_decls)
 
     def to_click(self) -> Callable[[click.Command], click.Command]:
@@ -78,8 +186,15 @@ class ListOption(AbstractOption):
             multiple=True,
         )
 
+    def to_widgets(self) -> Generator[Widget, None, None]:
+        raise NotImplementedError("No widget for ListOption.")
+
 
 class PathOption(AbstractOption):
+    """
+    A text input with path validation and autocomplete features.
+    """
+
     def __init__(
         self,
         name: str,
@@ -91,13 +206,36 @@ class PathOption(AbstractOption):
         dir_okay: bool = True,
         resolve_path: bool = False,
         path_type: type | None = Path,
+        default: str | None = None,
+        placeholder: str | None = None,
     ) -> None:
+        """
+        Args:
+            name (str): A unique name for this option. Must be a valid
+                HTML/CSS id and a valid CLI option name (without the `--` prefix).
+                e.g., "port", "header"
+            description (str): Help text for this option.
+            label (str | None): For GUI options, a human-friendly label for this option.
+            short_decls (Sequence[str] | None): For CLI options, a list of short aliases
+                (including the `-` prefix) for this option (e.g., ["-p"]).
+            exists (bool): *Validation* Set True if the path must already exist.
+            file_okay (bool): *Validation* Set True if the path may be a file.
+            dir_okay (bool):  *Validation* Set True if the path may be a directory.
+            resolve_path (bool): For CLI Options, set True for the returned path to be
+                resolved.
+            path_type (type): For CLI Options, define a type for the returned path
+                (usually str or pathlib.Path).
+            default (str): The default path.
+            placeholder (str): For GUI options, the placeholder text for the input.
+        """
         super().__init__(name, description, label=label, short_decls=short_decls)
         self.exists = exists
         self.file_okay = file_okay
         self.dir_okay = dir_okay
         self.resolve_path = resolve_path
         self.path_type = path_type
+        self.default = default
+        self.placeholder = placeholder
 
     def to_click(self) -> Callable[[click.Command], click.Command]:
         return click.option(
@@ -113,34 +251,126 @@ class PathOption(AbstractOption):
             ),
         )
 
+    def to_widgets(self) -> Generator[Widget, None, None]:
+        yield NoFocusLabel(f"{self.label}:", classes="input_label")
+        yield PathInput(
+            value=self.default or "",
+            placeholder=self.placeholder or "",
+            id=self.name,
+            file_okay=self.file_okay,
+            dir_okay=self.dir_okay,
+            must_exist=self.exists,
+            tab_advances_focus=True,
+        )
+
 
 class SelectOption(AbstractOption):
     def __init__(
         self,
         name: str,
         description: str,
-        choices: list[str],
+        choices: list[str] | list[tuple[str, str]],
         label: str | None = None,
         short_decls: list[str] | None = None,
+        default: str | None = None,
     ) -> None:
         super().__init__(name, description, label=label, short_decls=short_decls)
         self.choices = choices
+        self.default = default
+        """
+        Args:
+            name (str): A unique name for this option. Must be a valid
+                HTML/CSS id and a valid CLI option name (without the `--` prefix).
+                e.g., "port", "header"
+            description (str): Help text for this option.
+            choices (list[str] | list[tuple[str, str]]): A list of values or list of
+                (label, value) pairs for the user to select from.
+            label (str | None): For GUI options, a human-friendly label for this option.
+            short_decls (Sequence[str] | None): For CLI options, a list of short aliases
+                (including the `-` prefix) for this option (e.g., ["-p"]).
+            default (str | None): The default value for this option.
+        """
 
     def to_click(self) -> Callable[[click.Command], click.Command]:
+        choices: list[str] = []
+        for choice in self.choices:
+            if isinstance(choice, str):
+                choices.append(choice)
+            else:
+                choices.append(choice[0])
         return click.option(
             f"--{self.name}",
             *self.short_decls,
             help=self.description,
-            type=click.Choice(choices=self.choices, case_sensitive=False),
+            type=click.Choice(choices=choices, case_sensitive=False),
+        )
+
+    def to_widgets(self) -> Generator[Widget, None, None]:
+        choices: list[tuple[str, str]] = []
+        for choice in self.choices:
+            if isinstance(choice, str):
+                choices.append((choice, choice))
+            else:
+                choices.append(choice)
+        yield NoFocusLabel(f"{self.label}:", classes="select_label")
+        yield Select(
+            options=choices,
+            id=self.name,
+            value=self.default,
+            allow_blank=False,
         )
 
 
 class FlagOption(AbstractOption):
+    """
+    A boolean option, defaults to False.
+    """
+
     def to_click(self) -> Callable[[click.Command], click.Command]:
         return click.option(
             f"--{self.name}", *self.short_decls, help=self.description, is_flag=True
         )
 
+    def to_widgets(self) -> Generator[Widget, None, None]:
+        yield NoFocusLabel(f"{self.label}:", classes="switch_label")
+        yield Switch(id=self.name)
+
 
 HarlequinAdapterOption = AbstractOption
-HarlequinCopyOption = AbstractOption
+
+
+class HarlequinCopyFormat:
+    """
+    A file format for data export that is supported by the Adapter.
+    """
+
+    name: str
+    label: str
+    extensions: Sequence[str]
+    options: Sequence[HarlequinAdapterOption]
+
+    def __init__(
+        self,
+        name: str,
+        label: str | None = None,
+        extensions: Sequence[str] | None = None,
+        options: Sequence[HarlequinAdapterOption] | None = None,
+    ) -> None:
+        """
+        Args:
+            name (str): A unique, internal name for this format. E.g., 'csv'
+            label (str | None): A user-facing name for this format. E.g., 'CSV'
+            extensions (Sequence[str] | None): A seq of file extensions to associate
+                with this format. Should include the leading period. e.g.,
+                (".csv", ".tsv").
+            options (Sequence[HarlequinAdapterOption] | None): A list of options for
+                configuring copy operations for this format. e.g.,
+                [FlagOption(name="Header", description="Include header row?")]
+        """
+        self.name = name
+        self.label = label or name.replace("_", " ").replace("_", " ").capitalize()
+        self.extensions = extensions or tuple()
+        self.extensions = [
+            ext if ext.startswith(".") else f".{ext}" for ext in self.extensions
+        ]
+        self.options = options or tuple()